+++ MatrixSDKTests/MatrixSDKTestsData.m
@@ -17,8 +17,10 @@
 
 #import "MatrixSDKTestsData.h"
 
+#import "MXSDKOptions.h"
 #import "MXRestClient.h"
 #import "MXError.h"
+#import "MXNoStore.h"
 
 // Do not bother with retain cycles warnings in tests
 #pragma clang diagnostic push
@@ -917,6 +919,92 @@ onUnrecognizedCertificateBlock:(MXHTTPClientOnUnrecognizedCertificate)onUnrecogn
     }];
 }
 
+- (void)loginUserOnANewDevice:(XCTestCase*)testCase
+                  credentials:(MXCredentials*)credentials
+                 withPassword:(NSString*)password
+               sessionToLogout:(MXSession*)sessionToLogout
+              newSessionStore:(id<MXStore>)newSessionStore
+              startNewSession:(BOOL)startNewSession
+                          e2e:(BOOL)e2e
+                   onComplete:(void (^)(MXSession *newSession))onComplete
+{
+    if (!credentials && sessionToLogout)
+    {
+        credentials = sessionToLogout.credentials;
+    }
+    
+    if (sessionToLogout)
+    {
+        [sessionToLogout logout:^{
+            [sessionToLogout close];
+            
+            [self loginUserOnANewDevice:testCase credentials:credentials withPassword:password sessionToLogout:nil newSessionStore:newSessionStore startNewSession:startNewSession e2e:e2e onComplete:onComplete];
+            
+        } failure:^(NSError *error) {
+            [self breakTestCase:testCase reason:@"Cannot logout %@. Error: %@", sessionToLogout.myUserId, error];
+        }];
+        return;
+    }
+    
+    MXRestClient *mxRestClient = [[MXRestClient alloc] initWithHomeServer:credentials.homeServer
+                                        andOnUnrecognizedCertificateBlock:nil];
+    [self retain:mxRestClient];
+    
+    [mxRestClient loginWithLoginType:kMXLoginFlowTypePassword username:credentials.userId password:password success:^(MXCredentials *credentials2) {
+        
+        MXRestClient *mxRestClient2 = [[MXRestClient alloc] initWithCredentials:credentials2 andOnUnrecognizedCertificateBlock:nil];
+        [self retain:mxRestClient2];
+        
+        MXSession *newSession = [[MXSession alloc] initWithMatrixRestClient:mxRestClient2];
+        [self retain:newSession];
+        
+        if (!newSessionStore)
+        {
+            if (startNewSession)
+            {
+                [MXSDKOptions sharedInstance].enableCryptoWhenStartingMXSession = e2e;
+                [newSession start:^{
+                    [MXSDKOptions sharedInstance].enableCryptoWhenStartingMXSession = NO;
+                    
+                    onComplete(newSession);
+                    
+                } failure:^(NSError *error) {
+                    [self breakTestCase:testCase reason:@"Cannot start the session - error: %@", error];
+                }];
+            }
+            else
+            {
+                onComplete(newSession);
+            }
+            return;
+        }
+        
+        [newSession setStore:newSessionStore success:^{
+            if (startNewSession)
+            {
+                [MXSDKOptions sharedInstance].enableCryptoWhenStartingMXSession = e2e;
+                [newSession start:^{
+                    [MXSDKOptions sharedInstance].enableCryptoWhenStartingMXSession = NO;
+                    
+                    onComplete(newSession);
+                    
+                } failure:^(NSError *error) {
+                    [self breakTestCase:testCase reason:@"Cannot start the session - error: %@", error];
+                }];
+            }
+            else
+            {
+                onComplete(newSession);
+            }
+            
+        } failure:^(NSError *error) {
+            [self breakTestCase:testCase reason:@"Cannot open the store - error: %@", error];
+        }];
+        
+    } failure:^(NSError *error) {
+        [self breakTestCase:testCase reason:@"Cannot log %@ in again. Error: %@", credentials.userId , error];
+    }];
+}
 
 #pragma mark Reference keeping
 - (void)retain:(NSObject*)object