+++ CHANGES.rst
@@ -5,6 +5,9 @@ Changes to be released in next version
  * 
 
 🙌 Improvements
+ * MXDehydrationService: Support full rehydration feature (vector-im/element-ios/issues/1117).
+ * MXSDKOptions: Add wellknownDomainUrl to customise the domain for wellknown (vector-im/element-ios/issues/#4489).
+ * MXSession: Refresh homeserverWellknown on every start.
  * VoIP: Implement bridged version for call transfers.
  * VoIP: Implement MXiOSAudioOutputRouter.